LPS-55886 Move base classes to portal-test-internal
