+++ portal-impl/test/unit/com/liferay/portal/kernel/search/SearchResultUtilMBMessageTest.java
@@ -15,8 +15,8 @@
 package com.liferay.portal.kernel.search;
 
 import com.liferay.portal.kernel.comment.Comment;
-import com.liferay.portal.kernel.search.test.BaseSearchResultUtilTestCase;
-import com.liferay.portal.kernel.search.test.SearchTestUtil;
+import com.liferay.portal.search.test.BaseSearchResultUtilTestCase;
+import com.liferay.portal.search.test.SearchTestUtil;
 import com.liferay.portlet.asset.AssetRendererFactoryRegistryUtil;
 import com.liferay.portlet.messageboards.model.MBMessage;
 import com.liferay.registry.collections.ServiceTrackerCollections;
