+++ portal-impl/test/unit/com/liferay/portal/kernel/search/SearchResultUtilTest.java
@@ -14,10 +14,10 @@
 
 package com.liferay.portal.kernel.search;
 
-import com.liferay.portal.kernel.search.test.BaseSearchResultUtilTestCase;
-import com.liferay.portal.kernel.search.test.SearchTestUtil;
 import com.liferay.portal.kernel.test.util.RandomTestUtil;
 import com.liferay.portal.kernel.util.StringPool;
+import com.liferay.portal.search.test.BaseSearchResultUtilTestCase;
+import com.liferay.portal.search.test.SearchTestUtil;
 import com.liferay.portlet.asset.AssetRendererFactoryRegistryUtil;
 import com.liferay.registry.collections.ServiceTrackerCollections;
