+++ portal-test-internal/src/com/liferay/portal/search/test/SearchTestUtil.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.portal.kernel.search.test;
+package com.liferay.portal.search.test;
 
 import com.liferay.portal.kernel.search.Document;
 import com.liferay.portal.kernel.search.DocumentHelper;
